Adds @context to toplevel @context instead of to values
